+++ workers/ohsome_quality_analyst/cli/cli.py
@@ -199,19 +199,70 @@ def create_report(
 
 
 @cli.command("create-all-indicators")
+@click.option(
+    "--dataset-name",
+    "-d",
+    required=True,
+    type=click.Choice(
+        DATASETS.keys(),
+        case_sensitive=True,
+    ),
+    help=("Choose a dataset containing geometries."),
+)
+@click.option(
+    "--indicator-name",
+    "-i",
+    type=click.Choice(
+        load_metadata("indicators").keys(),
+        case_sensitive=True,
+    ),
+    help="Choose an indicator,valid indicators are specified in definitions.py .",
+    default=None,
+)
+@click.option(
+    "--layer-name",
+    "-l",
+    type=click.Choice(
+        list(load_layer_definitions().keys()),
+        case_sensitive=True,
+    ),
+    help=(
+        "Choose a layer. This defines which OSM features will be considered "
+        "in the quality analysis."
+    ),
+    default=None,
+)
 @cli_option(options.force)
-def create_all_indicators(force: bool):
-    """Create all indicators for all OQT regions."""
+def create_all_indicators(
+    dataset_name: str,
+    indicator_name: str,
+    layer_name: str,
+    force: bool,
+):
+    """Create all Indicators for all features of the given dataset.
+
+    The default is to create all Indicator/Layer combinations for all features of the
+    given dataset. This can be restricted to one Indicator type and/or one Layer
+    definition by providing the corresponding options.
+    """
     click.echo(
-        "This command will calculate all indicators for all OQT regions "
-        + "and may take a while to complete."
+        "This command will calculate all indicators for all features of the given "
+        + "dataset. This may take a while to complete."
     )
     if force:
         click.echo(
-            "The argument 'force' will update the indicator results in the database."
+            "The argument 'force' will overwrite existing indicator results in the "
+            + "database."
         )
     click.confirm("Do you want to continue?", abort=True)
-    asyncio.run(oqt.create_all_indicators(force=force))
+    asyncio.run(
+        oqt.create_all_indicators(
+            dataset_name,
+            indicator_name=indicator_name,
+            layer_name=layer_name,
+            force=force,
+        )
+    )
 
 
 if __name__ == "__main__":